+++ bootstrap.php
@@ -12,10 +12,17 @@
 
 namespace Flagrow\ImageUpload;
 
+use Flagrow\ImageUpload\Providers\StorageServiceProvider;
+use Flarum\Foundation\Application;
 use Illuminate\Contracts\Events\Dispatcher;
 
-return function (Dispatcher $events) {
+return function (Dispatcher $events, Application $app) {
+
+    // register the event listeners
     $events->subscribe(Listeners\AddClientAssets::class);
     $events->subscribe(Listeners\LoadSettingsFromDatabase::class);
     $events->subscribe(Listeners\AddUploadsApi::class);
+
+    // register the service provider
+    $app->register(StorageServiceProvider::class);
 };